Add Admin and his privileges

+++ BulgarianDestinations/Areas/Admin/Controllers/OrderController.cs
@@ -1,9 +1,14 @@
 ﻿using BulgarianDestinations.Core.Contracts;
 using BulgarianDestinations.Core.Services;
+using BulgarianDestinations.Infrastructure.Data.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using static BulgarianDestinations.Core.Constants.RoleConstants;
 
-namespace BulgarianDestinations.Controllers
+namespace BulgarianDestinations.Areas.Admin.Controllers
 {
+    [Area("Admin")]
+    [Authorize(Roles = AdminRole)]
     public class OrderController : Controller
     {
         private readonly IOrderService orderService;
@@ -14,15 +19,19 @@ namespace BulgarianDestinations.Controllers
             orderService = _orderService;
         }
         [HttpGet]
-        public async Task<IActionResult> All()
+        public IActionResult All()
         {
-            var model = await orderService.All();
+            var model = orderService.All();
             return View(model);
         }
 
         [HttpGet]
         public async Task<IActionResult> Details(int id)
         {
+            if (await orderService.Exists(id) == false)
+            {
+                return BadRequest();
+            }
             var model = await orderService.OrderInformation(id);
             return View(model);
         }
@@ -30,6 +39,10 @@ namespace BulgarianDestinations.Controllers
         [HttpGet]
         public async Task<IActionResult> Delete(int id)
         {
+            if (await orderService.Exists(id) == false)
+            {
+                return BadRequest();
+            }
 
             await orderService.DeleteOrder(id);
